+++ packages/proxy/schema/index.ts
@@ -5,12 +5,11 @@ import type {
   MessageRole,
   ModelParams,
 } from "@braintrust/core/typespecs";
+import { AvailableModels, ModelFormat } from "./models";
 
 export * from "./secrets";
+export * from "./models";
 
-export type PromptInputType = "completion" | "chat";
-
-export type ModelFormat = "openai" | "anthropic" | "google" | "js";
 export const ModelEndpointType = [
   "openai",
   "azure",
@@ -26,14 +25,6 @@ export const ModelEndpointType = [
   "js",
 ] as const;
 export type ModelEndpointType = (typeof ModelEndpointType)[number];
-
-export interface ModelSpec {
-  format: ModelFormat;
-  flavor: PromptInputType;
-  multimodal?: boolean;
-  input_cost_per_token?: number;
-  output_cost_per_token?: number;
-}
 
 export const MessageTypes: { [name in ModelFormat]: MessageRole[] } = {
   openai: ["system", "user", "assistant" /*, "function" */],
@@ -141,508 +132,6 @@ export const modelProviderHasTools: {
   anthropic: false,
   google: false,
   js: false,
-};
-
-export const AvailableModels: { [name: string]: ModelSpec } = {
-  "gpt-3.5-turbo": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000005,
-    output_cost_per_token: 0.0000015,
-  },
-  "gpt-35-turbo": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000005,
-    output_cost_per_token: 0.0000015,
-  },
-  "gpt-3.5-turbo-0125": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000005,
-    output_cost_per_token: 0.0000015,
-  },
-  "gpt-3.5-turbo-1106": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.000001,
-    output_cost_per_token: 0.000002,
-  },
-  "gpt-3.5-turbo-16k": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.000003,
-    output_cost_per_token: 0.000004,
-  },
-  "gpt-35-turbo-16k": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.000003,
-    output_cost_per_token: 0.000004,
-  },
-  "gpt-4": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00003,
-    output_cost_per_token: 0.00006,
-  },
-  "gpt-4o": {
-    format: "openai",
-    flavor: "chat",
-    multimodal: true,
-    input_cost_per_token: 0.000005,
-    output_cost_per_token: 0.000015,
-  },
-  "gpt-4o-2024-05-13": {
-    format: "openai",
-    flavor: "chat",
-    multimodal: true,
-    input_cost_per_token: 0.000005,
-    output_cost_per_token: 0.000015,
-  },
-  "gpt-4-32k": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00006,
-    output_cost_per_token: 0.00012,
-  },
-  "gpt-4-vision-preview": {
-    format: "openai",
-    flavor: "chat",
-    multimodal: true,
-    input_cost_per_token: 0.00001,
-    output_cost_per_token: 0.00003,
-  },
-  "gpt-4-1106-vision-preview": {
-    format: "openai",
-    flavor: "chat",
-    multimodal: true,
-    input_cost_per_token: 0.00001,
-    output_cost_per_token: 0.00003,
-  },
-  "gpt-4-turbo": {
-    format: "openai",
-    flavor: "chat",
-    multimodal: true,
-    input_cost_per_token: 0.00001,
-    output_cost_per_token: 0.00003,
-  },
-  "gpt-4-turbo-2024-04-09": {
-    format: "openai",
-    flavor: "chat",
-    multimodal: true,
-    input_cost_per_token: 0.00001,
-    output_cost_per_token: 0.00003,
-  },
-  "gpt-4-turbo-preview": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00001,
-    output_cost_per_token: 0.00003,
-  },
-  "gpt-4-0125-preview": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00001,
-    output_cost_per_token: 0.00003,
-  },
-  "gpt-4-1106-preview": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00001,
-    output_cost_per_token: 0.00003,
-  },
-  "gpt-3.5-turbo-0613": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000015,
-    output_cost_per_token: 0.000002,
-  },
-  "gpt-3.5-turbo-16k-0613": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.000003,
-    output_cost_per_token: 0.000004,
-  },
-  "gpt-3.5-turbo-0301": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000015,
-    output_cost_per_token: 0.000002,
-  },
-  "gpt-4-0613": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00003,
-    output_cost_per_token: 0.00006,
-  },
-  "gpt-4-32k-0613": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00006,
-    output_cost_per_token: 0.00012,
-  },
-  "gpt-4-0314": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00003,
-    output_cost_per_token: 0.00006,
-  },
-  "gpt-4-32k-0314": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00006,
-    output_cost_per_token: 0.00012,
-  },
-  "gpt-3.5-turbo-instruct": {
-    format: "openai",
-    flavor: "completion",
-    input_cost_per_token: 0.0000015,
-    output_cost_per_token: 0.000002,
-  },
-  "gpt-3.5-turbo-instruct-0914": {
-    format: "openai",
-    flavor: "completion",
-    input_cost_per_token: 0.0000015,
-    output_cost_per_token: 0.000002,
-  },
-  "text-davinci-003": {
-    format: "openai",
-    flavor: "completion",
-    input_cost_per_token: 0.000002,
-    output_cost_per_token: 0.000002,
-  },
-  "claude-2": {
-    format: "anthropic",
-    flavor: "chat",
-    input_cost_per_token: 0.000008,
-    output_cost_per_token: 0.000024,
-  },
-  "claude-instant-1": {
-    format: "anthropic",
-    flavor: "chat",
-    input_cost_per_token: 0.0000008,
-    output_cost_per_token: 0.0000024,
-  },
-  "claude-2.0": {
-    format: "anthropic",
-    flavor: "chat",
-    input_cost_per_token: 0.000008,
-    output_cost_per_token: 0.000024,
-  },
-  "claude-2.1": {
-    format: "anthropic",
-    flavor: "chat",
-    input_cost_per_token: 0.000008,
-    output_cost_per_token: 0.000024,
-  },
-  "claude-instant-1.2": {
-    format: "anthropic",
-    flavor: "chat",
-    input_cost_per_token: 0.0000008,
-    output_cost_per_token: 0.0000024,
-  },
-  "claude-3-opus-20240229": {
-    format: "anthropic",
-    flavor: "chat",
-    multimodal: true,
-    input_cost_per_token: 0.000015,
-    output_cost_per_token: 0.000075,
-  },
-  "claude-3-sonnet-20240229": {
-    format: "anthropic",
-    flavor: "chat",
-    multimodal: true,
-    input_cost_per_token: 0.000003,
-    output_cost_per_token: 0.000015,
-  },
-  "claude-3-haiku-20240307": {
-    format: "anthropic",
-    flavor: "chat",
-    multimodal: true,
-    input_cost_per_token: 0.00000025,
-    output_cost_per_token: 0.00000125,
-  },
-  "anthropic.claude-3-opus-20240229-v1:0": {
-    format: "anthropic",
-    flavor: "chat",
-    multimodal: true,
-    input_cost_per_token: 0.000015,
-    output_cost_per_token: 0.000075,
-  },
-  "anthropic.claude-3-haiku-20240307-v1:0": {
-    format: "anthropic",
-    flavor: "chat",
-    multimodal: true,
-    input_cost_per_token: 0.00000025,
-    output_cost_per_token: 0.00000125,
-  },
-  "anthropic.claude-3-sonnet-20240229-v1:0": {
-    format: "anthropic",
-    flavor: "chat",
-    multimodal: true,
-    input_cost_per_token: 0.000003,
-    output_cost_per_token: 0.000015,
-  },
-  "meta/llama-2-70b-chat": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00000065,
-    output_cost_per_token: 0.00000275,
-  },
-  "llama-2-70b-chat": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.000001,
-    output_cost_per_token: 0.000001,
-  },
-  "llama-2-13b-chat": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00000014,
-    output_cost_per_token: 0.00000056,
-  },
-  "llama3-8b-8192": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000001,
-    output_cost_per_token: 0.0000001,
-  },
-  "llama3-70b-8192": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00000064,
-    output_cost_per_token: 0.0000008,
-  },
-  "llama2-70b-4096": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000007,
-    output_cost_per_token: 0.0000008,
-  },
-  "codellama-34b-instruct": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00000035,
-    output_cost_per_token: 0.0000014,
-  },
-  "mistral-7b-instruct": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00000007,
-    output_cost_per_token: 0.00000028,
-  },
-  "mixtral-8x7b-instruct": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00000007,
-    output_cost_per_token: 0.00000028,
-  },
-  "mixtral-8x22b-instruct": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.000001,
-    output_cost_per_token: 0.000001,
-  },
-  "mixtral-8x7b-32768": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00000027,
-    output_cost_per_token: 0.00000027,
-  },
-  "gemma-7b-it": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000001,
-    output_cost_per_token: 0.0000001,
-  },
-  "mistralai/Mistral-7B-Instruct-v0.1": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000002,
-    output_cost_per_token: 0.0000002,
-  },
-  "mistralai/mixtral-8x7b-32kseqlen": {
-    format: "openai",
-    flavor: "completion",
-    input_cost_per_token: 0.00000006,
-    output_cost_per_token: 0.00000006,
-  },
-  "mistralai/Mixtral-8x7B-Instruct-v0.1": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000006,
-    output_cost_per_token: 0.0000006,
-  },
-  "mistralai/Mixtral-8x7B-Instruct-v0.1-json": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000006,
-    output_cost_per_token: 0.0000006,
-  },
-  "mistralai/Mixtral-8x22B": {
-    format: "openai",
-    flavor: "completion",
-    input_cost_per_token: 0.00000108,
-    output_cost_per_token: 0.00000108,
-  },
-  "mistralai/Mixtral-8x22B-Instruct-v0.1": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000012,
-    output_cost_per_token: 0.0000012,
-  },
-  "meta-llama/Llama-2-70b-chat-hf": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000009,
-    output_cost_per_token: 0.0000009,
-  },
-  "meta-llama/Meta-Llama-3-70B": {
-    format: "openai",
-    flavor: "completion",
-    input_cost_per_token: 0.0000009,
-    output_cost_per_token: 0.0000009,
-  },
-  "meta-llama/Llama-3-70b-chat-hf": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000009,
-    output_cost_per_token: 0.0000009,
-  },
-  "meta-llama/Llama-3-8b-hf": {
-    format: "openai",
-    flavor: "completion",
-    input_cost_per_token: 0.0000002,
-    output_cost_per_token: 0.0000002,
-  },
-  "meta-llama/Llama-3-8b-chat-hf": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000002,
-    output_cost_per_token: 0.0000002,
-  },
-  "NousResearch/Nous-Hermes-2-Yi-34B": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000008,
-    output_cost_per_token: 0.0000008,
-  },
-  "deepseek-ai/deepseek-coder-33b-instruct": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000008,
-    output_cost_per_token: 0.0000008,
-  },
-  "llama-3-8b-instruct": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000002,
-    output_cost_per_token: 0.0000002,
-  },
-  "llama-3-70b-instruct": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.000001,
-    output_cost_per_token: 0.000001,
-  },
-  "codellama-70b-instruct": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000007,
-    output_cost_per_token: 0.0000028,
-  },
-  mistral: {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0,
-    output_cost_per_token: 0.0,
-  },
-  "mistral-tiny": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00000015,
-    output_cost_per_token: 0.00000046,
-  },
-  "mistral-small": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.000002,
-    output_cost_per_token: 0.000006,
-  },
-  "mistral-medium": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000027,
-    output_cost_per_token: 0.0000081,
-  },
-  "openhermes-2-mistral-7b": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000002,
-    output_cost_per_token: 0.0000002,
-  },
-  "openhermes-2.5-mistral-7b": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000002,
-    output_cost_per_token: 0.0000002,
-  },
-  "pplx-7b-chat": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.00000007,
-    output_cost_per_token: 0.00000028,
-  },
-  "pplx-70b-chat": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000007,
-    output_cost_per_token: 0.0000028,
-  },
-  "pplx-7b-online": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0000002,
-    output_cost_per_token: 0.0000002,
-  },
-  "pplx-70b-online": {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.000001,
-    output_cost_per_token: 0.000001,
-  },
-  phi: {
-    format: "openai",
-    flavor: "chat",
-    input_cost_per_token: 0.0,
-    output_cost_per_token: 0.0,
-  },
-  "gemini-1.5-flash-latest": {
-    format: "google",
-    flavor: "chat",
-    input_cost_per_token: 0.0000007,
-    output_cost_per_token: 0.0000021,
-  },
-  "gemini-1.5-pro-latest": {
-    format: "google",
-    flavor: "chat",
-    input_cost_per_token: 0.000007,
-    output_cost_per_token: 0.000021,
-  },
-  "gemini-1.0-pro": {
-    format: "google",
-    flavor: "chat",
-    input_cost_per_token: 0.0000005,
-    output_cost_per_token: 0.0000015,
-  },
-  "gemini-pro": {
-    format: "google",
-    flavor: "chat",
-    input_cost_per_token: 0.00000025,
-    output_cost_per_token: 0.0000005,
-  },
-  "text-block": { format: "js", flavor: "completion" },
 };
 
 export const DefaultEndpointTypes: {
